requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
supabase==2.3.0
PyPDF2==3.0.0
python-docx==1.1.0
requests==2.31.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
